Address @michael-mclawhorn's comments
